benchmark/requirements.txt
# Graph Summarization Benchmark Requirements
# ==========================================

# Core Data Processing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Graph Processing  
networkx>=2.6.0
python-igraph>=0.9.6

# Performance and Parallel Processing
numba>=0.54.0
joblib>=1.0.0

# System Monitoring
psutil>=5.8.0
memory-profiler>=0.58.0

# Progress and Visualization
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration Management
PyYAML>=5.4.0
configparser>=5.0.0

# Data Serialization
# pickle5>=0.0.11  # Not needed for Python 3.8+
cloudpickle>=1.6.0

# Utilities and Testing
pytest>=6.2.0
pytest-benchmark>=3.4.0
click>=8.0.0

# Jupyter (for analysis)
jupyter>=1.0.0
ipykernel>=6.0.0

# Report Generation
markdown>=3.3.0
jinja2>=3.0.0
tabulate>=0.8.0

# Hash Functions (for MinHash implementation)
mmh3>=3.0.0
datasketch>=1.5.0

# Graph File I/O
python-louvain>=0.15
# community>=0.13.0  # Có conflict, skip for now

# Statistical Analysis
scikit-learn>=1.0.0
statsmodels>=0.12.0

# Optional: GPU acceleration (if available)
# cupy-cuda11x>=9.0.0  # uncomment if CUDA available
# rapids-cudf>=21.06   # uncomment if Rapids available 